`qc2.py`: Ensure integers are returned when computing widths/heights

+++ spinalcordtoolbox/reports/qc2.py
@@ -1132,8 +1132,8 @@ def get_max_axial_radius(img):
     default = 15
     heights = [np.max(np.where(slice)[0]) - np.min(np.where(slice)[0]) if np.sum(slice) > 0 else 0 for slice in img.data]
     widths = [np.max(np.where(slice)[1]) - np.min(np.where(slice)[1]) if np.sum(slice) > 0 else 0 for slice in img.data]
-    heights = [(h * dilation)//2 for h in heights]
-    widths = [(w * dilation)//2 for w in widths]
+    heights = [int((h * dilation)//2) for h in heights]
+    widths = [int((w * dilation)//2) for w in widths]
     return max(default, max(heights)), max(default, max(widths))
